refactor(demod): move ChannelDemod (renamed from PdDemod) + FFT_LEN / pixel_freq / freq_to_luminance / ycbcr_to_rgb to crate::demod (#85 B1, C6)
