+++ src/mode_pd.rs
@@ -8,216 +8,7 @@
 //! Channel layout: video.c lines 81-93. YUV→RGB matrix: video.c lines
 //! 446-451. See `NOTICE.md` for full attribution.
 
-use rustfft::{num_complex::Complex, FftPlanner};
-
-/// Map a demodulated FM frequency (Hz) to an 8-bit luminance value.
-///
-/// SSTV video lives in 1500–2300 Hz: 1500 Hz = black (0), 2300 Hz = white (255).
-/// Linear scaling: `lum = (freq - 1500) / (2300 - 1500) * 255`.
-/// Out-of-band frequencies are clamped. `hedr_shift_hz` shifts the band
-/// to compensate for radio mistuning detected at VIS time:
-/// `lum = (freq - 1500 - hedr_shift_hz) / 3.1372549`.
-///
-/// Translated from slowrx's `video.c:406` + `common.c:49-53`:
-/// `StoredLum[SampleNum] = clip((Freq - 1500 - HedrShift) / 3.1372549);`
-/// where `clip(a)` returns `(guchar)round(a)` clamped to \[0, 255\].
-/// The `round()` call means values like 127.7 map to 128, not 127.
-///
-/// `3.1372549 = (2300 - 1500) / 255`.
-#[must_use]
-pub(crate) fn freq_to_luminance(freq_hz: f64, hedr_shift_hz: f64) -> u8 {
-    let v = (freq_hz - 1500.0 - hedr_shift_hz) / 3.137_254_9;
-    // Round-to-nearest before casting, matching slowrx's `(guchar)round(a)`
-    // in `common.c::clip()`.
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    let lum = v.clamp(0.0, 255.0).round() as u8;
-    lum
-}
-
-/// Convert a single PD-family `[Y, Cr, Cb]` triple to `[R, G, B]`.
-///
-/// Translated from slowrx's `video.c:447-450` + `common.c:49-53`:
-/// ```text
-/// R = clip((100*Y + 140*Cr - 17850) / 100.0)
-/// G = clip((100*Y -  71*Cr -  33*Cb + 13260) / 100.0)
-/// B = clip((100*Y + 178*Cb - 22695) / 100.0)
-/// ```
-/// where `clip(a)` is `(guchar)round(a)` clamped to \[0, 255\] (`common.c:49-53`).
-///
-/// slowrx uses `/ 100.0` (float division), which produces a `double` that is
-/// then **rounded** by `clip()` before clamping. The previous implementation
-/// used integer division (`/ 100`), which **truncates toward zero** — this
-/// produced a 1-LSB darker bias on R and B channels for neutral grey and many
-/// other combinations. Fixed in round-2 audit Finding 1.
-#[must_use]
-#[doc(hidden)]
-pub fn ycbcr_to_rgb(y: u8, cr: u8, cb: u8) -> [u8; 3] {
-    let yi = f64::from(y);
-    let cri = f64::from(cr);
-    let cbi = f64::from(cb);
-    // Float divide then round, matching slowrx's `clip(double)` in common.c:49-53:
-    //   `return (guchar)round(a)` after clamping to [0, 255].
-    // i32 intermediate magnitudes are well within f64 precision (max 255*178=45390).
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    let r = ((100.0 * yi + 140.0 * cri - 17_850.0) / 100.0)
-        .clamp(0.0, 255.0)
-        .round() as u8;
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    let g = ((100.0 * yi - 71.0 * cri - 33.0 * cbi + 13_260.0) / 100.0)
-        .clamp(0.0, 255.0)
-        .round() as u8;
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    let b = ((100.0 * yi + 178.0 * cbi - 22_695.0) / 100.0)
-        .clamp(0.0, 255.0)
-        .round() as u8;
-    [r, g, b]
-}
-
-/// FFT length used for per-pixel demod. Matches slowrx's bin spacing:
-/// `1024/11025 Hz` ≈ 10.77 Hz/bin — 4× finer than slowrx's `1024/44100 Hz`
-/// = 43.07 Hz/bin (deliberate 0.3.3 divergence; see
-/// `docs/intentional-deviations.md`).
-pub(crate) const FFT_LEN: usize = crate::snr::FFT_LEN;
-
-/// Per-pixel demod context: holds an FFT plan + reusable buffers + the
-/// adaptive Hann-window bank. Construct once per decoder; reuse for many
-/// [`PdDemod::pixel_freq`] calls.
-pub(crate) struct PdDemod {
-    fft: std::sync::Arc<dyn rustfft::Fft<f32>>,
-    hann_bank: crate::demod::HannBank,
-    fft_buf: Vec<Complex<f32>>,
-    scratch: Vec<Complex<f32>>,
-}
-
-impl PdDemod {
-    pub fn new() -> Self {
-        let mut planner = FftPlanner::<f32>::new();
-        let fft = planner.plan_fft_forward(FFT_LEN);
-        let scratch_len = fft.get_inplace_scratch_len();
-        Self {
-            fft,
-            hann_bank: crate::demod::HannBank::new(),
-            fft_buf: vec![Complex { re: 0.0, im: 0.0 }; FFT_LEN],
-            scratch: vec![Complex { re: 0.0, im: 0.0 }; scratch_len.max(FFT_LEN)],
-        }
-    }
-
-    /// Estimate the dominant tone frequency in a Hann-windowed FFT
-    /// centered near `center_sample`. `win_idx` selects the Hann length
-    /// from [`crate::demod::HANN_LENS`]; the FFT length stays fixed at
-    /// [`FFT_LEN`] (= 1024), with leading/trailing zero-pad. Out-of-bounds
-    /// `audio` reads as silence. `hedr_shift_hz` shifts the peak-search
-    /// range from `[1500, 2300]` Hz to `[1500 + hedr, 2300 + hedr]` Hz to
-    /// follow a mistuned radio's pixel band.
-    ///
-    /// Translated from slowrx `video.c:369-395` (windowed FFT + bin
-    /// search + Gaussian interp).
-    #[allow(
-        clippy::cast_precision_loss,
-        clippy::cast_possible_truncation,
-        clippy::cast_sign_loss,
-        clippy::cast_possible_wrap
-    )]
-    pub fn pixel_freq(
-        &mut self,
-        audio: &[f32],
-        center_sample: i64,
-        hedr_shift_hz: f64,
-        win_idx: usize,
-    ) -> f64 {
-        let win_idx = win_idx.min(crate::demod::HANN_LENS.len() - 1);
-        let hann = self.hann_bank.get(win_idx);
-        let win_len = hann.len();
-
-        // Zero-fill, then apply the windowed support centered on
-        // `center_sample`. slowrx (`video.c:369-375`) writes the
-        // `WinLength` windowed samples into the FIRST `WinLength` bins
-        // of the FFT input; the magnitude spectrum is invariant to that
-        // offset (just a phase rotation).
-        for c in &mut self.fft_buf {
-            *c = Complex { re: 0.0, im: 0.0 };
-        }
-        let half = (win_len as i64) / 2;
-        for (i, (&w, dst)) in hann.iter().zip(self.fft_buf.iter_mut()).enumerate() {
-            let idx = center_sample - half + i as i64;
-            let s = if idx >= 0 && (idx as usize) < audio.len() {
-                audio[idx as usize]
-            } else {
-                0.0
-            };
-            *dst = Complex { re: s * w, im: 0.0 };
-        }
-
-        self.fft
-            .process_with_scratch(&mut self.fft_buf, &mut self.scratch[..]);
-
-        // Search peak in bins corresponding to (1500+HedrShift)..(2300+HedrShift) Hz.
-        // Use slowrx-equivalent truncation (not `.round()`) via `crate::dsp::get_bin`.
-        // See `crate::dsp::get_bin` for rationale.
-        let bin_for = |hz: f64| -> usize {
-            crate::dsp::get_bin(hz, FFT_LEN, crate::resample::WORKING_SAMPLE_RATE_HZ)
-        };
-        let lo = bin_for(1500.0 + hedr_shift_hz).saturating_sub(1).max(1);
-        let hi = bin_for(2300.0 + hedr_shift_hz)
-            .saturating_add(1)
-            .min(FFT_LEN / 2 - 1);
-
-        let mut max_bin = lo;
-        let mut max_p = crate::dsp::power(self.fft_buf[lo]);
-        for (k, &c) in self.fft_buf.iter().enumerate().take(hi + 1).skip(lo + 1) {
-            let p = crate::dsp::power(c);
-            if p > max_p {
-                max_p = p;
-                max_bin = k;
-            }
-        }
-
-        // Boundary clip + Gaussian-log peak interpolation (slowrx video.c:389-398).
-        //
-        // slowrx's guard (`video.c:390`):
-        //   if (MaxBin > GetBin(1500+HedrShift) - 1 &&
-        //       MaxBin < GetBin(2300+HedrShift) + 1) { interpolate }
-        //   else { Freq = (MaxBin > GetBin(1900+HedrShift)) ? 2300 : 1500 + HedrShift; }
-        //
-        // `lo` = GetBin(1500+hedr) - 1 and `hi` = GetBin(2300+hedr) + 1 (as above),
-        // so the guard translates to: `max_bin > lo && max_bin < hi`. When the peak
-        // lands on one of the padded boundary bins, slowrx returns a hard-clipped
-        // value rather than interpolating into the neighbor noise bin (round-2 audit
-        // Finding 9).
-        let mid_bin = bin_for(1900.0 + hedr_shift_hz);
-        if max_bin <= lo || max_bin >= hi {
-            // Clip to band edge (slowrx video.c:397).
-            let clipped_hz = if max_bin > mid_bin {
-                2300.0 + hedr_shift_hz
-            } else {
-                1500.0 + hedr_shift_hz
-            };
-            return clipped_hz;
-        }
-
-        // Freq_bin = MaxBin + log(P[k+1]/P[k-1]) / (2 * log(P[k]^2 / (P[k+1] * P[k-1])))
-        let p_prev = crate::dsp::power(self.fft_buf[max_bin - 1]);
-        let p_curr = max_p;
-        let p_next = crate::dsp::power(self.fft_buf[max_bin + 1]);
-
-        // If any neighbor power is non-positive, skip interpolation
-        // (log of zero blows up). slowrx falls back to a clipped centre.
-        let interp_ok = p_prev > 0.0 && p_curr > 0.0 && p_next > 0.0;
-        let freq_bin = if interp_ok {
-            let num = (p_next / p_prev).ln();
-            let denom = 2.0 * (p_curr * p_curr / (p_next * p_prev)).ln();
-            if denom.abs() > 1e-12 {
-                (max_bin as f64) + num / denom
-            } else {
-                max_bin as f64
-            }
-        } else {
-            max_bin as f64
-        };
-
-        freq_bin * f64::from(crate::resample::WORKING_SAMPLE_RATE_HZ) / (FFT_LEN as f64)
-    }
-}
+use crate::demod::{freq_to_luminance, ycbcr_to_rgb, ChannelDemod, FFT_LEN, SNR_REESTIMATE_STRIDE};
 
 /// Distance, in working-rate samples, between successive per-pixel
 /// FFTs. slowrx takes an FFT every 6 samples at `44_100` Hz
@@ -229,11 +20,6 @@ impl PdDemod {
 /// difference vs slowrx's stride=6 is negligible on offline-batch
 /// decoding at `11_025` Hz.
 const PIXEL_FFT_STRIDE: i64 = 1;
-
-/// Distance, in working-rate samples, between successive SNR
-/// re-estimates. slowrx re-estimates every 256 samples at `44_100` Hz
-/// (`video.c:343`); scaled to `11_025` Hz that's 64.
-pub(crate) const SNR_REESTIMATE_STRIDE: i64 = 64;
 
 /// Decode one PD radio frame (`Y(odd)`/`Cr`/`Cb`/`Y(even)`) into two image
 /// rows of `image`. Translated from slowrx `video.c:259-486`.
@@ -291,7 +77,7 @@ pub(crate) fn decode_pd_line_pair(
     pair_seconds: f64,
     rate_hz: f64,
     image: &mut crate::image::SstvImage,
-    demod: &mut PdDemod,
+    demod: &mut ChannelDemod,
     snr_est: &mut crate::snr::SnrEstimator,
     hedr_shift_hz: f64,
 ) {
@@ -398,7 +184,7 @@ pub(crate) fn decode_one_channel_into(
     skip_samples: i64,
     time_offset_seconds: f64,
     rate_hz: f64,
-    demod: &mut PdDemod,
+    demod: &mut ChannelDemod,
     snr_est: &mut crate::snr::SnrEstimator,
     hedr_shift_hz: f64,
 ) {
@@ -537,8 +323,6 @@ pub(crate) fn decode_one_channel_into(
     clippy::cast_possible_wrap
 )]
 mod tests {
-    use super::*;
-
     /// Verify that with `septr_seconds = 0` the `chan_starts_sec` formula
     /// gives the same values as the pre-#25 direct formula (numeric equivalence).
     /// This confirms the field is a V2 expansion that is a no-op for PD modes.
@@ -579,240 +363,5 @@ mod tests {
                 );
             }
         }
-    }
-
-    #[test]
-    fn freq_1500_is_black() {
-        assert_eq!(freq_to_luminance(1500.0, 0.0), 0);
-    }
-
-    #[test]
-    fn freq_2300_is_white() {
-        assert_eq!(freq_to_luminance(2300.0, 0.0), 255);
-    }
-
-    #[test]
-    fn freq_below_band_clamps_to_zero() {
-        assert_eq!(freq_to_luminance(1000.0, 0.0), 0);
-        assert_eq!(freq_to_luminance(0.0, 0.0), 0);
-    }
-
-    #[test]
-    fn freq_above_band_clamps_to_max() {
-        assert_eq!(freq_to_luminance(3000.0, 0.0), 255);
-    }
-
-    #[test]
-    fn freq_midband_is_midgrey() {
-        // 1900 Hz: v = (1900 - 1500) / 3.1372549 ≈ 127.5 → rounds to 128
-        let v = freq_to_luminance(1900.0, 0.0);
-        assert!(
-            (i32::from(v) - 128).abs() <= 1,
-            "midband should be ~128 after rounding, got {v}"
-        );
-    }
-
-    #[test]
-    fn freq_to_luminance_rounds_to_nearest_not_truncates() {
-        // Slowrx uses `(guchar)round(a)` in common.c::clip(), not truncation.
-        // Solve (freq - 1500) / 3.1372549 = 127.7 → freq ≈ 1900.6294 Hz.
-        // round(127.7) = 128, not 127.
-        let freq = 1500.0 + 127.7 * 3.137_254_9; // ≈ 1900.629...
-        assert_eq!(
-            freq_to_luminance(freq, 0.0),
-            128,
-            "127.7 should round to 128, not truncate to 127"
-        );
-    }
-
-    #[test]
-    fn freq_to_luminance_with_hedr_shift_scales_band() {
-        // With +50 Hz HedrShift, the band shifts to 1550..2350.
-        // freq=1550 should be black, freq=2350 should be white.
-        assert_eq!(freq_to_luminance(1550.0, 50.0), 0);
-        assert_eq!(freq_to_luminance(2350.0, 50.0), 255);
-        // 1500 (would be black at zero shift) is now sub-band, still clamps to 0.
-        assert_eq!(freq_to_luminance(1500.0, 50.0), 0);
-        // 1950 with +50 shift is the new midband ≈ same as 1900 with 0 shift.
-        let a = freq_to_luminance(1950.0, 50.0);
-        let b = freq_to_luminance(1900.0, 0.0);
-        assert!(
-            i32::from(a).abs_diff(i32::from(b)) <= 1,
-            "shifted midband {a} vs unshifted {b}"
-        );
-    }
-
-    #[test]
-    fn ycbcr_neutral_grey_is_grey() {
-        // Y=128, Cr=128, Cb=128 → neutral grey.
-        // Exact values (slowrx float-divide + round):
-        //   R = (100*128 + 140*128 - 17850) / 100.0 = 12870/100.0 = 128.7 → round → 129
-        //   G = (100*128 -  71*128 -  33*128 + 13260)/100.0 = 12748/100.0 = 127.48 → round → 127
-        //   B = (100*128 + 178*128 - 22695)/100.0 = 12889/100.0 = 128.89 → round → 129
-        let rgb = ycbcr_to_rgb(128, 128, 128);
-        for ch in &rgb {
-            assert!((i32::from(*ch) - 128).abs() <= 2, "got {rgb:?}");
-        }
-    }
-
-    /// Verify round-to-nearest parity with slowrx's `clip()` (`common.c:49-53`).
-    ///
-    /// slowrx uses `clip((100*Y + 140*Cr - 17850) / 100.0)` where `clip` calls
-    /// `round()`. Y=128, Cr=128, Cb=128 produces:
-    ///   R numerator = 12870 → 128.70 → round → **129** (not 128 from integer division)
-    ///   B numerator = 12889 → 128.89 → round → **129** (not 128 from integer division)
-    ///
-    /// This test would fail with the old integer-division implementation.
-    #[test]
-    fn ycbcr_rounds_to_nearest_matching_slowrx_clip() {
-        let [r, g, b] = ycbcr_to_rgb(128, 128, 128);
-        assert_eq!(
-            r, 129,
-            "R should be 129 (round(128.70)), not 128 (truncate)"
-        );
-        assert_eq!(g, 127, "G should be 127 (round(127.48))");
-        assert_eq!(
-            b, 129,
-            "B should be 129 (round(128.89)), not 128 (truncate)"
-        );
-    }
-
-    #[test]
-    fn ycbcr_pure_red() {
-        // Roughly: max Cr, mid Y, mid Cb → strong red.
-        let rgb = ycbcr_to_rgb(76, 255, 85);
-        assert!(rgb[0] > 200, "red channel should dominate, got {rgb:?}");
-        assert!(rgb[2] < 100);
-    }
-
-    use crate::resample::WORKING_SAMPLE_RATE_HZ;
-    use std::f64::consts::PI;
-
-    fn synth_tone(freq_hz: f64, secs: f64) -> Vec<f32> {
-        let n = (secs * f64::from(WORKING_SAMPLE_RATE_HZ)).round() as usize;
-        (0..n)
-            .map(|i| {
-                let t = (i as f64) / f64::from(WORKING_SAMPLE_RATE_HZ);
-                (2.0 * PI * freq_hz * t).sin() as f32
-            })
-            .collect()
-    }
-
-    /// Default `win_idx` for tests: index 4 (Hann length 64 at our
-    /// `11_025` Hz working rate), slowrx's middle-of-the-band default.
-    const DEFAULT_WIN_IDX: usize = 4;
-
-    #[test]
-    fn pdfft_recovers_known_tone_within_5hz() {
-        let mut d = PdDemod::new();
-        // 100 ms of pure tone at 1900 Hz; ample for FFT_LEN=1024.
-        let audio = synth_tone(1900.0, 0.100);
-        let center = (audio.len() / 2) as i64;
-        let est = d.pixel_freq(&audio, center, 0.0, DEFAULT_WIN_IDX);
-        assert!((est - 1900.0).abs() < 5.0, "expected ≈1900, got {est}");
-    }
-
-    #[test]
-    fn pdfft_recovers_band_edges() {
-        let mut d = PdDemod::new();
-        for &f in &[1500.0_f64, 1700.0, 2100.0, 2300.0] {
-            let audio = synth_tone(f, 0.100);
-            let center = (audio.len() / 2) as i64;
-            let est = d.pixel_freq(&audio, center, 0.0, DEFAULT_WIN_IDX);
-            assert!(
-                (est - f).abs() < 8.0,
-                "f={f} estimate={est} (band edge precision)"
-            );
-        }
-    }
-
-    #[test]
-    fn pdfft_returns_finite_for_silence() {
-        let mut d = PdDemod::new();
-        let audio = vec![0.0_f32; 1024];
-        let est = d.pixel_freq(&audio, 512, 0.0, DEFAULT_WIN_IDX);
-        assert!(est.is_finite(), "got {est}");
-        // Silence has no peak; the search expands by ±1 bin around the
-        // 1500-2300 band, and bin width is ~43 Hz, so the fallback may
-        // land within ~50 Hz of either edge.
-        assert!((1450.0..=2350.0).contains(&est), "out of band: {est}");
-    }
-
-    #[test]
-    fn pixel_freq_with_hedr_shift() {
-        // A 1950 Hz tone with hedr_shift=+50 should yield the same luminance
-        // as a 1900 Hz tone with hedr_shift=0, since the search band shifts.
-        let mut d = PdDemod::new();
-        let audio = synth_tone(1950.0, 0.100);
-        let center = (audio.len() / 2) as i64;
-        let est_shifted = d.pixel_freq(&audio, center, 50.0, DEFAULT_WIN_IDX);
-        let est_unshifted_baseline = {
-            let baseline = synth_tone(1900.0, 0.100);
-            d.pixel_freq(&baseline, (baseline.len() / 2) as i64, 0.0, DEFAULT_WIN_IDX)
-        };
-        // Tone-frequency itself is recovered correctly; what matters is that
-        // the shifted estimator finds 1950, the unshifted finds 1900, and
-        // mapping each to luminance gives ≈ same value.
-        assert!((est_shifted - 1950.0).abs() < 5.0, "got {est_shifted}");
-        let lum_shifted = freq_to_luminance(est_shifted, 50.0);
-        let lum_unshifted = freq_to_luminance(est_unshifted_baseline, 0.0);
-        assert!(
-            i32::from(lum_shifted).abs_diff(i32::from(lum_unshifted)) <= 2,
-            "lum_shifted={lum_shifted} lum_unshifted={lum_unshifted}"
-        );
-    }
-
-    #[test]
-    fn pixel_freq_clamps_out_of_range_win_idx() {
-        // Defensive: an out-of-range win_idx must not panic.
-        let mut d = PdDemod::new();
-        let audio = synth_tone(1900.0, 0.100);
-        let center = (audio.len() / 2) as i64;
-        let est = d.pixel_freq(&audio, center, 0.0, 99);
-        // Falls back to the longest window (idx 6) and still recovers the tone.
-        assert!((est - 1900.0).abs() < 10.0, "clamp recover got {est}");
-    }
-
-    #[test]
-    fn pixel_freq_short_window_still_recovers_tone() {
-        // The shortest window (idx 0, length 12) is intended for high-SNR
-        // signals. With a clean synthetic tone it should still localize the
-        // peak inside the video band, just with coarser precision than a
-        // long window.
-        let mut d = PdDemod::new();
-        let audio = synth_tone(1900.0, 0.100);
-        let center = (audio.len() / 2) as i64;
-        let est = d.pixel_freq(&audio, center, 0.0, 0);
-        assert!(
-            (1500.0..=2300.0).contains(&est),
-            "short-window estimate {est} out of video band"
-        );
-    }
-
-    /// Verify that a tone below the search band clips to 1500 Hz (round-2 audit
-    /// Finding 9 — boundary clip matching slowrx `video.c:395-397`).
-    ///
-    /// At 1480 Hz the FFT peak lands on the padded boundary bin (lo) or below.
-    /// slowrx's guard:
-    ///   `if (MaxBin > GetBin(1500+HedrShift)-1 && MaxBin < GetBin(2300+HedrShift)+1)`
-    /// fails, so it returns `1500 + HedrShift` (the lower clip).
-    /// The estimate must be ≤ 1500 Hz and within one bin-width (~43 Hz) of 1500 Hz.
-    #[test]
-    fn pixel_freq_clips_below_band_to_1500hz() {
-        let mut d = PdDemod::new();
-        // 1480 Hz is 1 bin-width (~43 Hz) below the search range start (1500 Hz).
-        // The FFT peak for this tone will land at or below the padded lo boundary.
-        let audio = synth_tone(1480.0, 0.100);
-        let center = (audio.len() / 2) as i64;
-        let est = d.pixel_freq(&audio, center, 0.0, DEFAULT_WIN_IDX);
-        // Must be clipped to ≈1500 Hz, not freely interpolated toward a DC neighbor.
-        assert!(
-            est <= 1500.0 + 50.0,
-            "below-band tone should clip to ≈1500 Hz, got {est}"
-        );
-        assert!(
-            est >= 1400.0,
-            "clip floor should be near 1500 Hz, got {est}"
-        );
     }
 }
